test(helpdesk): refactor fixtures and unskip chat attachment tests
