+++ tests/e2e/helpdesk/chats/attachment/test_sync_attachment.py
@@ -8,7 +8,6 @@ from mpt_api_client.resources.helpdesk.chat_attachments import ChatAttachment
 pytestmark = [pytest.mark.flaky]
 
 
-@pytest.mark.skip(reason="Unskip after MPT-19124 completed")
 def test_list_chat_attachments(chat_attachments_service, created_chat_attachment):
     result = chat_attachments_service.fetch_page(limit=1)
 
@@ -16,20 +15,17 @@ def test_list_chat_attachments(chat_attachments_service, created_chat_attachment
     assert all(isinstance(attachment, ChatAttachment) for attachment in result)
 
 
-@pytest.mark.skip(reason="Unskip after MPT-19124 completed")  # noqa: AAA01
-def test_create_chat_attachment(created_chat_attachment, chat_attachment_data):
+def test_create_chat_attachment(created_chat_attachment, chat_attachment_data):  # noqa: AAA01
     assert created_chat_attachment.id is not None
     assert created_chat_attachment.to_dict().get("name") == chat_attachment_data["name"]
 
 
-@pytest.mark.skip(reason="Unskip after MPT-19124 completed")
 def test_get_chat_attachment(chat_attachments_service, created_chat_attachment):
     result = chat_attachments_service.get(created_chat_attachment.id)
 
     assert result.id == created_chat_attachment.id
 
 
-@pytest.mark.skip(reason="Unskip after MPT-19124 completed")
 def test_update_chat_attachment(chat_attachments_service, created_chat_attachment, short_uuid):
     updated_name = f"e2e updated attachment - {short_uuid}"
 
@@ -42,18 +38,18 @@ def test_update_chat_attachment(chat_attachments_service, created_chat_attachmen
     assert result.to_dict().get("name") == updated_name
 
 
-@pytest.mark.skip(reason="Unskip after MPT-19124 completed")
 def test_download_chat_attachment(chat_attachments_service, created_chat_attachment):
     result = chat_attachments_service.download(created_chat_attachment.id, accept="application/pdf")
 
     assert result.file_contents is not None
 
 
-@pytest.mark.skip(reason="Unskip after MPT-19124 completed")  # noqa: AAA01
 def test_delete_chat_attachment(chat_attachments_service, chat_attachment_data, pdf_fd):
     created = chat_attachments_service.create(chat_attachment_data, file=pdf_fd)
 
-    chat_attachments_service.delete(created.id)
+    result = chat_attachments_service.delete(created.id)
+
+    assert result is None
 
 
 def test_get_chat_attachment_not_found(chat_attachments_service, invalid_chat_attachment_id):
