+++ tests/e2e/helpdesk/chats/messages/conftest.py
@@ -7,13 +7,13 @@ from tests.e2e.helper import (
 
 
 @pytest.fixture
-def chat_messages_service(mpt_ops, chat_id):
-    return mpt_ops.helpdesk.chats.messages(chat_id)
+def chat_messages_service(mpt_ops, created_chat):
+    return mpt_ops.helpdesk.chats.messages(created_chat.id)
 
 
 @pytest.fixture
-def async_chat_messages_service(async_mpt_ops, chat_id):
-    return async_mpt_ops.helpdesk.chats.messages(chat_id)
+def async_chat_messages_service(async_mpt_ops, created_chat):
+    return async_mpt_ops.helpdesk.chats.messages(created_chat.id)
 
 
 @pytest.fixture
