+++ tests/e2e/helpdesk/conftest.py
@@ -48,3 +48,53 @@ async def async_created_disabled_queue(async_mpt_ops, async_created_queue):
     assert result.status == "Disabled"
 
     return result
+
+
+@pytest.fixture(scope="session")
+def invalid_case_id():
+    return "CAS-0000-0000"
+
+
+@pytest.fixture
+def case_data(created_queue):
+    return {
+        "queue": {"id": created_queue.id},
+        "chat": {"lastMessage": {"content": "E2E testing!!!"}},
+    }
+
+
+@pytest.fixture
+def created_case(mpt_ops, case_data):
+    return mpt_ops.helpdesk.cases.create(case_data)
+
+
+@pytest.fixture
+def queried_case(mpt_ops, created_case):
+    return mpt_ops.helpdesk.cases.query(created_case.id, {"queryPrompt": "More details needed"})
+
+
+@pytest.fixture
+def processed_case(mpt_ops, queried_case):
+    return mpt_ops.helpdesk.cases.process(queried_case.id)
+
+
+@pytest.fixture
+def created_chat(created_case):
+    return created_case.chat
+
+
+@pytest.fixture
+async def async_created_case(async_mpt_ops, case_data):
+    return await async_mpt_ops.helpdesk.cases.create(case_data)
+
+
+@pytest.fixture
+async def async_queried_case(async_mpt_ops, async_created_case):
+    return await async_mpt_ops.helpdesk.cases.query(
+        async_created_case.id, {"queryPrompt": "More details needed"}
+    )
+
+
+@pytest.fixture
+async def async_processed_case(async_mpt_ops, async_queried_case):
+    return await async_mpt_ops.helpdesk.cases.process(async_queried_case.id)
